provide example flyer; move controller to proper folder
