+++ bluebottle/utils/templates/utils/share_flyer.hbs
@@ -9,7 +9,10 @@
     <div class="modal-fullscreen-content">
         <span {{action 'closeModal'}} class="modal-fullscreen-close"></span>
         <h4>{% blocktrans %}Share a Project flyer{% endblocktrans %}</h4>
-        <big>{% blocktrans %}Share this project with your partners and/or invite them to join (view example){% endblocktrans %}</big>
+        <big>{% trans "Share this project with your partners and/or invite them to join" %}
+             (<a href="/utils/share_flyer?projectid={{unbound model.projectid}}" target="_blank"> {% trans "view example" %}</a>)
+
+        </big>
 
          <label for="contactName">{% trans "Recipients name" %}* </label>
         {{input id="contactName" valueBinding='share_name' type="text" name="share_name" maxlength="254" placeholder=App.translatables.placeholders.share_name classBinding="errorList.share_name.length:error"}}
